move runCommandHandleLines
